backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.11.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
stripe==11.4.1
httpx==0.28.1
jinja2==3.1.5
python-dotenv==1.0.1
